% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cosinor.R
\docType{data}
\name{cosinor_batch}
\alias{cosinor_batch}
\title{Calculate cosinor measures from HBA actigraphy data
Always refer to https://rpubs.com/Jonathon_Pye for full explanation and package dependencies
After setting your working directory, assign your .csv's to 'files' and run the loop.
_cosinor and _cosinor_plots.csv will be saved in your working directory}
\format{
An object of class \code{list} of length 0.
}
\usage{
cosinor_batch
}
\description{
Calculate cosinor measures from HBA actigraphy data
Always refer to https://rpubs.com/Jonathon_Pye for full explanation and package dependencies
After setting your working directory, assign your .csv's to 'files' and run the loop.
_cosinor and _cosinor_plots.csv will be saved in your working directory
}
\examples{




process_cosinor()
}
\keyword{actigraphy,}
\keyword{cosinor}
\keyword{process,}
