% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actig_batch_process.R,
%   R/unique_dates_and_missingness.R
\docType{data}
\name{files}
\alias{files}
\title{Process and impute HBA actigraphy data}
\format{
An object of class \code{character} of length 0.

An object of class \code{character} of length 0.
}
\usage{
files

files
}
\arguments{
\item{files}{list of .csv actigraphy files in your WD}
}
\description{
After setting your working directory, assign your .csv's to 'files' and run the loop

After setting your working directory, assign your .csv's to 'files' and run the loop as shown in example
}
\examples{
files(.csv) # list of files in wd.




actig_batch_process()
unique_dates_and_missingness <- do.call(rbind, lapply(files, unique_dates_and_missingness))



unique_dates_and_missingness()
}
\keyword{actigraphy}
\keyword{actigraphy,}
\keyword{impute,}
\keyword{length}
\keyword{missing,}
\keyword{process,}
