% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_nonparametric.R
\name{process_nonparametric}
\alias{process_nonparametric}
\title{Calculate nonparametric measures for HBA actigraphy recordings
Always refer to https://rpubs.com/Jonathon_Pye for full explanation and package dependencies
After setting your working directory, assign your .csv's to 'files' and run the loop.
npar_results will be saved in your working directory and the header names will be printed in the console}
\usage{
process_nonparametric(files, filename)
}
\description{
Calculate nonparametric measures for HBA actigraphy recordings
Always refer to https://rpubs.com/Jonathon_Pye for full explanation and package dependencies
After setting your working directory, assign your .csv's to 'files' and run the loop.
npar_results will be saved in your working directory and the header names will be printed in the console
}
\examples{




process_nonparametric()
}
\keyword{actigraphy,}
\keyword{nonparametric}
\keyword{process,}
